small updates

+++ coordinates.cpp
@@ -1,4 +1,5 @@
 #include "coordinates.h"
+#include <cstddef>
 #include <limits>
 #include <iostream>
 
@@ -28,7 +29,7 @@ void coordinates::settings_reader(std::ifstream& ifile){
 
 //output of coordinates
 void coordinates::output_writer_partial(std::ofstream& ofile, int n_pos, std::vector<bool> atoms){
-    for(int i=0;i<siti.size();i++){
+    for(size_t i=0;i<siti.size();i++){
         if(atoms[i]){
             ofile<<specie_chimica[i]<<" "<<siti[i][0]<<" "<<siti[i][1]<<" "<<siti[i][2]<<std::endl;
         }
@@ -39,7 +40,7 @@ void coordinates::output_writer_partial(std::ofstream& ofile, int n_pos, std::ve
 void coordinates::output_writer(std::ofstream& ofile, int n_pos, double time){
     ofile<<n_pos<<std::endl;
     ofile << " time of the simulation: " << time << std::endl;
-    for(int i=0;i<siti.size();i++){
+    for(size_t i=0;i<siti.size();i++){
             ofile<<specie_chimica[i]<<" "<<siti[i][0]<<" "<<siti[i][1]<<" "<<siti[i][2]<<std::endl;
     }
 }
